+++ poc/sisaq/src/main/java/br/com/mineradora/resource/AquisicaoResource.java
@@ -1,6 +1,8 @@
 package br.com.mineradora.resource;
 
 
+import java.math.BigInteger;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
@@ -31,7 +33,7 @@ public class AquisicaoResource {
     }
 	
 	@RequestMapping(value = "{id}" , method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<?> findById(@PathVariable("id") final Long id) {
+    public ResponseEntity<?> findById(@PathVariable("id") final BigInteger id) {
         return ResponseEntity.ok(this.aquisicaoService.findById(id));
     }
